+++ CHANGELOG.md
@@ -26,6 +26,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   #268).
 - Removed '_external_load', '_secundary_borefield_load', '_building_load' from Borefield.py (issue #268).
 - Vfr in FluidData now returns also a value if vfr = None based on the mfr.
+- Splitted Results class into ResultsHourly and ResultsMonthly (issue #281).
 
 ## Fixed